+++ docs/conf.py
@@ -20,12 +20,8 @@
 #
 import os
 import sys
+from importlib.metadata import PackageNotFoundError, version
 from shutil import copyfile
-
-try:
-    from importlib.metadata import PackageNotFoundError, version
-except ImportError:
-    from importlib_metadata import PackageNotFoundError, version
 
 sys.path.insert(0, os.path.abspath("../src/"))
 
@@ -79,7 +75,7 @@ try:
     # The full version, including alpha/beta/rc tags.
     release = version("reuse")
 except PackageNotFoundError:
-    release = "1.1.2"
+    release = "2.0.0"
 
 # The short X.Y.Z version.
 version = ".".join(release.split(".")[:3])